Uninstall Forgejo
It seems to be a bit more resource intensive than my little node can handle.

It also does not support actions (yet) and woodpecker is just another process to run, and a new
workflow configuration to learn. I like the idea of a decentralized forge, but that is also not
finished (yet).  Many of these same problems are also true for Codeberg, so I'm going to stick with
Github for now.

The network effect is hard to overcome.  I can't meaningfully collaborate with others if it requires
them to create an account on my self-hosted forge?

I'm somewhat sympathetic to concerns about Github, but there's not really lock-in since I have the
complete history of my git repos locally.  I guess I could be concerned about using my code for
training AI, etc., but it's open source, so ...